optimized scripts and renamed distribute-poly.js to 0_poly_transfer.js
