+++ scripts/cli/util/init.js
@@ -1,24 +1,21 @@
-const { ApiPromise, WsProvider } = require("@polkadot/api");
-export { ApiPromise, WsProvider };
-const {Keyring} = require("@polkadot/keyring");
-export {Keyring};
+export const { ApiPromise, WsProvider } = require("@polkadot/api");
+export const { Keyring } = require("@polkadot/keyring");
 export const BN = require("bn.js");
 export const cli = require("command-line-args");
 export const cliProg = require("cli-progress");
 const childProc = require("child_process");
-export const colors = require('colors');
+export const colors = require("colors");
 
 export const fs = require("fs");
 export const path = require("path");
-
 
 // Helps track the size delta for
 export let current_storage_size = 0;
 
 // Updated by the CLI option
-export let STORAGE_DIR;
+let STORAGE_DIR;
 export let nonces = new Map();
-export let alice, bob, dave;
+export let entities = [];
 export let master_keys = [];
 export let signing_keys = [];
 export let claim_keys = [];
@@ -32,46 +29,116 @@ export let block_times = {};
 export let synced_block = 0;
 export let synced_block_ts = 0;
 
+// Amount to seed each key with
+export let transfer_amount = 10 * 10 ** 12;
 
-// const cli_opts = [
-//   {
-//     name: "accounts", // Number of transactions/accounts to use per step
-//     alias: "n",
-//     type: Number,
-//     defaultValue: 30
-//   },
-//   {
-//     name: "claim_accounts", // Number of transactions/accounts to use per step
-//     alias: "t",
-//     type: Number,
-//     defaultValue: 5
-//   },
-//   {
-//     name: "claims", // How many claims to add to the `ntxs` DIDs
-//     alias: "c",
-//     type: Number,
-//     defaultValue: 10
-//   },
-//   {
-//     name: "prepend", // Prepend for secretUrl for uniqueness
-//     alias: "p",
-//     type: String,
-//     defaultValue: ""
-//   },
-//   {
-//     name: "dir", // Substrate storage dir
-//     alias: "d",
-//     type: String,
-//     defaultValue: "/tmp/pmesh-primary-node"
-//   },
-//   {
-//     name: "fast", // Substrate storage dir
-//     alias: "f",
-//     type: Boolean,
-//     defaultValue: false
-//   }
-// ];
-//exports.PrintNearestStore = async function(api) {
+// Parse CLI args and compute tx count
+const opts = {
+  accounts: 5,
+  claim_accounts: 5,
+  claims: 5,
+  prepend: "demo",
+  fast: false,
+  dir: "/tmp/pmesh-primary-node"
+};
+
+// CLI args variables
+let n_accounts = opts.accounts;
+let n_claim_accounts = opts.claim_accounts;
+let n_claims = opts.claims;
+let prepend = opts.prepend;
+let fast = opts.fast;
+
+const keyring = new Keyring({ type: "sr25519" });
+STORAGE_DIR = opts.dir;
+const initial_storage_size = duDirSize(STORAGE_DIR);
+current_storage_size = initial_storage_size;
+
+// Initialization Main is used to generate all entities e.g (Alice, Bob, Dave)
+// and Master keys, Signing keys, and Claim keys.
+async function initMain(api) {
+  console.log(
+    `Welcome to Polymesh Stats Collector. Creating ${n_accounts} accounts and DIDs, with ${n_claims} claims per DID.`
+  );
+
+  console.log(
+    `Initial storage size (${STORAGE_DIR}): ${initial_storage_size / 1024}MB`
+  );
+
+  generateEntity(api, "Alice");
+  generateEntity(api, "Bob");
+  generateEntity(api, "Dave");
+
+  // Create `n_accounts` master key accounts
+  console.log("Generating Master Keys");
+  generateMasterKeys(api);
+
+  // Create `n_accounts` signing key accounts
+  console.log("Generating Signing Keys");
+  generateSigningKeys(api);
+
+  // Create `n_accounts` claim key accounts
+  console.log("Generating Claim Keys");
+  generateClaimKeys(api);
+
+  updateStorageSize(STORAGE_DIR);
+}
+
+let generateEntity = async function(api, name) {
+  let entity = keyring.addFromUri(`//${name}`, { name: `${name}` });
+  entities[name] = entity;
+  let entityRawNonce = await api.query.system.accountNonce(
+    entities[name].address
+  );
+  let entity_nonce = new BN(entityRawNonce.toString());
+  nonces.set(entities[name].address, entity_nonce);
+};
+
+const generateMasterKeys = async function(api) {
+  for (let i = 0; i < n_accounts; i++) {
+    master_keys.push(
+      keyring.addFromUri("//IssuerMK" + prepend + i.toString(), {
+        name: i.toString()
+      })
+    );
+    let accountRawNonce = await api.query.system.accountNonce(
+      master_keys[i].address
+    );
+    let account_nonce = new BN(accountRawNonce.toString());
+    nonces.set(master_keys[i].address, account_nonce);
+  }
+};
+
+const generateSigningKeys = async function(api) {
+  for (let i = 0; i < n_accounts; i++) {
+    signing_keys.push(
+      keyring.addFromUri("//IssuerSK" + prepend + i.toString(), {
+        name: i.toString()
+      })
+    );
+    let accountRawNonce = await api.query.system.accountNonce(
+      signing_keys[i].address
+    );
+    let account_nonce = new BN(accountRawNonce.toString());
+    nonces.set(signing_keys[i].address, account_nonce);
+  }
+};
+
+const generateClaimKeys = async function(api) {
+  for (let i = 0; i < n_claim_accounts; i++) {
+    claim_keys.push(
+      keyring.addFromUri("//ClaimIssuerMK" + prepend + i.toString(), {
+        name: i.toString()
+      })
+    );
+    let claimIssuerRawNonce = await api.query.system.accountNonce(
+      claim_keys[i].address
+    );
+    let account_nonce = new BN(claimIssuerRawNonce.toString());
+    nonces.set(claim_keys[i].address, account_nonce);
+  }
+};
+
 const blockTillPoolEmpty = async function(api) {
   let prev_block_pending = 0;
   let done_something = false;
@@ -87,7 +154,9 @@ const blockTillPoolEmpty = async function(api) {
           done_something = true;
         }
         let timestamp_extrinsic = block["block"]["extrinsics"][0];
-        let new_block_ts = parseInt(JSON.stringify(timestamp_extrinsic.raw["method"].args[0].raw));
+        let new_block_ts = parseInt(
+          JSON.stringify(timestamp_extrinsic.raw["method"].args[0].raw)
+        );
         block_times[i] = new_block_ts - synced_block_ts;
         synced_block_ts = new_block_ts;
         synced_block = i;
@@ -103,7 +172,7 @@ const blockTillPoolEmpty = async function(api) {
   while (!done) {
     await new Promise(resolve => setTimeout(resolve, 1000));
   }
-}
+};
 
 // Use the `du` command to obtain recursive directory size
 function duDirSize(dir) {
@@ -114,6 +183,7 @@ function duDirSize(dir) {
   return new Number(results[1]);
 }
 
+// Updating storage size
 function updateStorageSize(dir) {
   let new_storage_size = duDirSize(dir);
   current_storage_size = new_storage_size;
@@ -129,9 +199,8 @@ const createIdentities = async function(api, accounts, fast) {
     if (fast) {
       await api.tx.identity
         .registerDid([])
-        .signAndSend(accounts[i],
-          { nonce: nonces.get(accounts[i].address) });
-    } 
+        .signAndSend(accounts[i], { nonce: nonces.get(accounts[i].address) });
+    }
     nonces.set(accounts[i].address, nonces.get(accounts[i].address).addn(1));
   }
   await blockTillPoolEmpty(api);
@@ -140,8 +209,17 @@ const createIdentities = async function(api, accounts, fast) {
     dids.push(d.raw.asUnique);
   }
   return dids;
-}
+};
 
-export { duDirSize, updateStorageSize, blockTillPoolEmpty, createIdentities};
-
-
+export {
+  duDirSize,
+  updateStorageSize,
+  blockTillPoolEmpty,
+  createIdentities,
+  initMain,
+  n_claim_accounts,
+  n_accounts,
+  STORAGE_DIR,
+  initial_storage_size,
+  current_storage_size
+};
